+++ components/Header.tsx
@@ -1,23 +1,10 @@
-import Image from 'next/image';
+import KakaoLoginButton from './KakaoLoginButton';
 
 interface HeaderProps {
   accessToken: string;
 }
 
 export default function Header({ accessToken }: HeaderProps) {
-  function receiveKakaoAuthorizationCode() {
-    const url = `${process.env.NEXT_PUBLIC_KAKAO_HOST
-    }/oauth/authorize?client_id=${process.env.NEXT_PUBLIC_REST_API_KEY
-    }&redirect_uri=${process.env.NEXT_PUBLIC_REDIRECT_URI
-    }&response_type=code`;
-
-    window.location.href = url;
-  }
-
-  function handleKakaoLoginClick() {
-    receiveKakaoAuthorizationCode();
-  }
-
   return (
     <header className="flex items-center flex-row-reverse w-screen relative h-28 px-3">
       <div className="w-60 absolute top-1/2 left-1/2 -translate-y-1/2 -translate-x-1/2">
@@ -25,9 +12,7 @@ export default function Header({ accessToken }: HeaderProps) {
         <h1 className="text-right text-5xl font-semibold">Arabang</h1>
       </div>
       <div>
-        <button type="button" onClick={handleKakaoLoginClick}>
-          <Image src="/images/kakao_login_medium_narrow.png" alt="kakao-login" width={130} height={50} />
-        </button>
+        <KakaoLoginButton />
       </div>
     </header>
   );
